pyproject.toml
[tool.poetry]
name = "metrics"
version = "0.1.0"
description = ""
authors = ["Szymon Idziniak <szymon.idziniak@stxnext.pl>", "Bartosz Mielczarek <bartosz.mielczarek@stxnext.pl>"]

# TODO: add torch into poetry dependencies,
# so it will be easier to build container just with requirements.txt dump from poetry
[tool.poetry.dependencies]
python = "^3.10.8"
scipy = "^1.9.0"
tensorboard = "^2.9.1"
umap-learn = "^0.5.3"
matplotlib = "^3.5.2"
Pillow = "^9.2.0"
scikit-learn = "^1.1.2"
timm = "^0.6.7"
qdrant-client = "^0.11"
uvicorn = {extras = ["standard"], version = "^0.19"}
fastapi = {extras = ["uvicorn"], version = "^0.87.0"}
pytorch-metric-learning = "^1.5.2"
pandas = "^1.4.4"
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
fastapi-utils = "^0.2.1"
record-keeper = "^0.9.32"
loguru = "^0.6.0"
streamlit = "^1.15.0"
minio = "^7.1.12"
streamlit-image-select = "^0.3.0"
faiss-cpu = "^1.7"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
autoflake = "^1"
mypy = "^0.991"
notebook = "^6.4.12"
jupyterlab = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# TODO: add exclude folders for python tools
[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.pytest_cache
    | venv
    | .venv
    | migrations
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_glob = ["/data/*"]